move factory utility into the scripts folder
