+++ test/integration/authorization/authorization.test.js
@@ -2,7 +2,7 @@ const Lab = require('lab')
 const lab = exports.lab = Lab.script()
 
 const server = require('./../../../lib/server')
-const Factory = require('../../factory')
+const Factory = require('../../../scripts/factory')
 const BuildFor = require('./testBuilder')
 
 const organizationId = 'WONKA'
