+++ test/integration/authorization/users.test.js
@@ -3,7 +3,7 @@ const Lab = require('lab')
 const lab = exports.lab = Lab.script()
 
 const server = require('./../../../lib/server')
-const Factory = require('../../factory')
+const Factory = require('../../../scripts/factory')
 const BuildFor = require('./testBuilder')
 
 const utils = require('../../utils')
